# Speed up shape_key_apply_modifiers using buffers