@greisane 2022-09-05
Very nice, thank you for the info and benchmark. I had assumed the benefit to foreach_get was just avoiding repeated API calls.